Refactor chat context into ts
