+++ src/react-components/room/LoadingScreen.js
@@ -2,7 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import { LoadingScreenLayout } from "../layout/LoadingScreenLayout";
 import { Spinner } from "../misc/Spinner";
-import { useRandomMessageTransition } from "./useRandomMessageTransition";
+import { useRandomMessageTransition } from "./hooks/useRandomMessageTransition";
 export function LoadingScreen({ message, infoMessages }) {
   const infoMessage = useRandomMessageTransition(infoMessages);
   return (
