+++ src/react-components/room/MicSetupModalContainer.js
@@ -1,17 +1,17 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { MicSetupModal } from "./MicSetupModal";
-import { useMicrophoneStatus } from "./useMicrophoneStatus";
-import { useMicrophone } from "./useMicrophone";
-import { useSound } from "./useSound";
+import { useMicrophoneStatus } from "./hooks/useMicrophoneStatus";
+import { useMicrophone } from "./hooks/useMicrophone";
+import { useSound } from "./hooks/useSound";
 import { SOUND_SPEAKER_TONE } from "../../systems/sound-effects-system";
-import { useSpeakers } from "./useSpeakers";
+import { useSpeakers } from "./hooks/useSpeakers";
 import { useCallback } from "react";
 import { useState } from "react";
 import MediaDevicesManager from "../../utils/media-devices-manager";
 import { VolumeLevelBar } from "../misc/VolumeLevelBar";
 import styles from "./MicSetupModal.scss";
-import { useCan } from "./useCan";
+import { useCan } from "./hooks/useCan";
 
 export function MicSetupModalContainer({ scene, ...rest }) {
   const { isMicEnabled, permissionStatus } = useMicrophoneStatus(scene);
