+++ src/react-components/room/ObjectMenuContainer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { ObjectMenu, ObjectMenuButton } from "./ObjectMenu";
-import { useObjectList } from "./useObjectList";
+import { useObjectList } from "./hooks/useObjectList";
 import {
   usePinObject,
   useRemoveObject,
