+++ src/react-components/room/SharePopoverContainer.js
@@ -5,7 +5,7 @@ import { ReactComponent as DesktopIcon } from "../icons/Desktop.svg";
 import { ReactComponent as AvatarIcon } from "../icons/Avatar.svg";
 import { SharePopoverButton } from "./SharePopover";
 import { FormattedMessage } from "react-intl";
-import useAvatar from "./useAvatar";
+import useAvatar from "./hooks/useAvatar";
 import { MediaDevicesEvents, MediaDevices } from "../../utils/media-devices-utils";
 
 function useShare(scene, hubChannel) {
