+++ src/react-components/room/hooks/useAvatar.js
