+++ src/react-components/room/hooks/useSpeakers.js
@@ -1,5 +1,5 @@
 import { useState, useEffect, useCallback } from "react";
-import { MediaDevices, MediaDevicesEvents } from "../../utils/media-devices-utils";
+import { MediaDevices, MediaDevicesEvents } from "../../../utils/media-devices-utils";
 
 export function useSpeakers() {
   const mediaDevicesManager = APP.mediaDevicesManager;
